Refactoring testAlert
